fix schema for profile (height + weight)

+++ backend/models/schemas.py
@@ -152,8 +152,8 @@ class UserProfileResponse(BaseModel):
     user_id: str
     age: Optional[int] = None
     gender: Optional[str] = None
-    weight: Optional[int] = None
-    height: Optional[int] = None
+    weight: Optional[float] = None
+    height: Optional[float] = None
     hba1c: Optional[float] = None
     activity_level: Optional[str] = None
     diet_restrictions: Optional[str] = None
@@ -165,8 +165,8 @@ class UserProfileResponse(BaseModel):
 class UserProfileUpdateRequest(BaseModel):
     age: Optional[int] = None
     gender: Optional[str] = None
-    weight: Optional[int] = None
-    height: Optional[int] = None
+    weight: Optional[float] = None
+    height: Optional[float] = None
     hba1c: Optional[float] = None
     activity_level: Optional[str] = None
     diet_restrictions: Optional[str] = None
